Moved classes related to parameter from framework.py to options.py. Moved def make_split_provider() from dataloader.py to options.py.
