+++ lib/__init__.py
@@ -1,18 +1,16 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .options import check_train_options, check_test_options
-from .framework import create_model, set_params, print_parameters
+from .options import check_train_options, check_test_options, ParamStore
+from .framework import create_model
 from .metrics import set_eval
 from .logger import BaseLogger
 
 __all__ = [
             'check_train_options',
             'check_test_options',
-            'set_params',
-            'print_parameters',
+            'ParamStore',
             'create_model',
             'set_eval',
             'BaseLogger'
         ]
-
