+++ lib/component/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .dataloader import make_split_provider, create_dataloader, print_dataset_info
+from .dataloader import create_dataloader, print_dataset_info
 from .net import create_net
 from .criterion import set_criterion
 from .optimizer import set_optimizer
@@ -9,7 +9,6 @@ from .loss import create_loss_store
 from .likelihood import set_likelihood
 
 __all__ = [
-            'make_split_provider',
             'create_dataloader',
             'print_dataset_info',
             'create_net',
