+++ lib/framework.py
@@ -1,395 +1,19 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from __future__ import annotations
 from pathlib import Path
 import copy
 from abc import ABC, abstractmethod
 import pandas as pd
 import torch
 import torch.nn as nn
-import json
-from .component import make_split_provider, create_net
+from .component import create_net
 from .logger import BaseLogger
-from typing import List, Dict, Tuple, Union
-import argparse
+from typing import Dict, Tuple, Union
+from lib import ParamStore
 
 
 logger = BaseLogger.get_logger(__name__)
-
-
-class BaseParam:
-    """
-    Set up configure for traning or test.
-    Integrate args and parameters.
-    """
-    def __init__(self, args: argparse.Namespace) -> None:
-        """
-        Args:
-            args (argparse.Namespace): arguments
-        """
-        for _param, _arg in vars(args).items():
-            setattr(self, _param, _arg)
-
-
-class ParamMixin:
-    """
-    class for save and load parameters
-    """
-    def save_parameter(self, save_datetime_dir: str) -> None:
-        """
-        Save parameters.
-
-        Args:
-            save_datetime_dir (str): save_datetime_dir
-        """
-        no_save = [
-                    'isTrain',
-                    'device',  # Need str(self.device) if save
-                    'df_source',
-                    'datetime',
-                    'save_datetime_dir',
-                    'dataloader_params',
-                    'model_params',
-                    'train_conf_params',
-                    'test_conf_params'
-                    ]
-
-        saved = dict()
-        for _param, _arg in vars(self).items():
-            if _param not in no_save:
-                saved[_param] = _arg
-
-        # Save parameters
-        save_dir = Path(save_datetime_dir)
-        save_dir.mkdir(parents=True, exist_ok=True)
-        save_path = str(Path(save_dir, 'parameters.json'))
-        with open(save_path, 'w') as f:
-            json.dump(saved, f, indent=4)
-
-    def load_parameter(self, parameter_path: Path) -> Dict:
-        """
-        Return dictionalry of parameters at training.
-
-        Args:
-            parameter_path (str): path to parameter_path
-
-        Returns:
-            Dict: parameters at training
-        """
-        with open(parameter_path) as f:
-            parameters = json.load(f)
-        return parameters
-
-
-class ParamWidget(BaseParam, ParamMixin):
-    """
-    Class for a widget to inherit multiple classes simultaneously.
-    """
-    pass
-
-
-class TrainParam(ParamWidget):
-    """
-    Class for setting parameters for training.
-    """
-    def __init__(self, args: argparse.Namespace) -> None:
-        """
-        Args:
-            args (argparse.Namespace): arguments
-        """
-        super().__init__(args)
-
-        sp = make_split_provider(self.csvpath, self.task)
-        self.df_source = sp.df_source
-        self.input_list = sp.input_list
-        self.label_list = sp.label_list
-        self.mlp_num_inputs = len(self.input_list)
-        self.num_outputs_for_label = self._define_num_outputs_for_label(self.df_source, self.label_list, self.task)
-
-        if self.task == 'deepsurv':
-            self.period_name = sp.period_name
-
-    def _define_num_outputs_for_label(self, df_source: pd.DataFrame, label_list: List[str], task :str) -> Dict[str, int]:
-        """
-        Define the number of outputs for each label.
-
-        Args:
-            df_source (pd.DataFrame): DataFrame of csv
-            label_list (List[str]): label list
-            task: str
-
-        Returns:
-            Dict[str, int]: dictionary of the number of outputs for each label
-            eg.
-                classification:       _num_outputs_for_label = {label_A: 2, label_B: 3, ...}
-                regression, deepsurv: _num_outputs_for_label = {label_A: 1, label_B: 1, ...}
-                deepsurv:             _num_outputs_for_label = {label_A: 1}
-        """
-        if task == 'classification':
-            _num_outputs_for_label = {label_name: df_source[label_name].nunique() for label_name in label_list}
-        elif (task == 'regression') or (task == 'deepsurv'):
-            _num_outputs_for_label = {label_name: 1 for label_name in label_list}
-        else:
-            raise ValueError(f"Invalid task: {task}.")
-        return _num_outputs_for_label
-
-
-class TestParam(ParamWidget):
-    """
-    Class for setting parameters for test.
-    """
-    def __init__(self, args: argparse.Namespace) -> None:
-        """
-        Args:
-            args (argparse.Namespace): arguments
-        """
-        super().__init__(args)
-
-        # Load paramaters
-        parameters = self.load_parameter(self.parameter_path)
-
-        required_for_test = [
-                            'task',
-                            'model',
-                            'normalize_image',
-                            'in_channel',
-                            'vit_image_size',
-                            'mlp',
-                            'net',
-                            'input_list',  # should be used one at trainig
-                            'label_list',  # shoudl be used one at trainig
-                            'period_name',
-                            'mlp_num_inputs',
-                            'num_outputs_for_label'
-                            ]
-
-        for _param in required_for_test:
-            if _param in parameters:
-                setattr(self, _param, parameters[_param])
-
-        # No need at test
-        self.augmentation = 'no'
-        self.sampler = 'no'
-        self.pretrained = False
-
-        if self.mlp is not None:
-            setattr(self, 'scaler_path', str(Path(self.train_save_datetime_dir, 'scaler.pkl')))
-
-        sp = make_split_provider(self.csvpath, self.task)  # should be done after task is loaded.
-        self.df_source = sp.df_source
-
-        _splits_in_df_source = self.df_source['split'].unique().tolist()
-        self.test_splits = self._align_test_splits(self.test_splits, _splits_in_df_source)
-
-    def _align_test_splits(self, arg_test_splits: List[str], splits_in_df_source: List[str]) -> List[str]:
-        """
-        Align splits to be test.
-
-        Args:
-            arg_test_splits (List[str]): splits specified by args. Default is ['train', 'val', 'test']
-            splits_in_df_source (List[str]): splits includinf csv
-
-        Returns:
-            List[str]: splits for test
-
-        args_test_splits  = ['train', 'val', 'test'], ['val', 'test'], or ['test']
-        splits_in_df_source = ['train', 'val', 'test'], or ['test']
-        Smaller set of splits has priority.
-        """
-        if set(splits_in_df_source) < set(arg_test_splits):
-            _test_splits = splits_in_df_source  # maybe when external dataset
-        elif set(arg_test_splits) < set(splits_in_df_source):
-            _test_splits = arg_test_splits      # ['val', 'test'], or ['test']
-        else:
-            _test_splits = arg_test_splits
-        return _test_splits
-
-
-class ParamGroup:
-    """
-    Class to register parameter for groups.
-    """
-    dataloader = [
-                'task',
-                'isTrain',
-                'df_source',
-                'label_list',
-                'input_list',
-                'period_name',
-                'batch_size',
-                'test_batch_size',
-                'mlp',
-                'net',
-                'scaler_path',
-                'in_channel',
-                'normalize_image',
-                'augmentation',
-                'sampler'
-                ]
-
-    net = [
-            'mlp',
-            'net',
-            'num_outputs_for_label',
-            'mlp_num_inputs',
-            'in_channel',
-            'vit_image_size',
-            'pretrained',
-            'gpu_ids'
-            ]
-
-    model = \
-            net + \
-            [
-            'task',
-            'isTrain',
-            'criterion',
-            'device',
-            'optimizer',
-            'lr',
-            'label_list',
-            ]
-
-    train_conf = [
-                'epochs',
-                'save_weight_policy',
-                'save_datetime_dir'
-                ]
-
-    test_conf = [
-                'task',
-                'weight_dir',
-                'weight_paths',
-                'num_outputs_for_label',
-                'test_splits',
-                'save_datetime_dir'
-                ]
-
-
-class ParamStore:
-    """
-    Class to store parameters for each group.
-    """
-    @classmethod
-    def dispatch_params_by_group(cls, params: Union[TrainParam, TestParam], group_name: str) -> ParamStore:
-        """
-        Dispatch parameters depenidng on group.
-
-        Args:
-            params (Union[TrainParam, TestParam]): parameters
-            group_name (str): group
-
-        Returns:
-            ParamStore: class containing parameters for group
-        """
-        for param_name in getattr(ParamGroup, group_name):
-            if hasattr(params, param_name):
-                _arg = getattr(params, param_name)
-                setattr(cls, param_name, _arg)
-        return cls
-
-
-def set_params(args: argparse.Namespace) -> Union[TrainParam, TestParam]:
-    """
-    Set parameters depending on training or test and dispatch them.
-
-    Args:
-        args (argparse.Namespace): args
-
-    Returns:
-        Union[TrainParam, TestParam]: parameters
-    """
-    if args.isTrain:
-        params = TrainParam(args)
-        params.dataloader_params = ParamStore.dispatch_params_by_group(params, 'dataloader')
-        params.model_params = ParamStore.dispatch_params_by_group(params, 'model')
-        params.train_conf_params = ParamStore.dispatch_params_by_group(params, 'train_conf')
-        return params
-    else:
-        params = TestParam(args)
-        params.dataloader_params = ParamStore.dispatch_params_by_group(params, 'dataloader')
-        params.model_params = ParamStore.dispatch_params_by_group(params, 'model')
-        params.test_conf_params = ParamStore.dispatch_params_by_group(params, 'test_conf')
-        return params
-
-
-def print_parameters(params: Union[TrainParam, TestParam]) -> None:
-    """
-    Print parameters
-
-    Args:
-        params (Union[TrainParam, TestParam]): parameters
-    """
-
-    # --- print order ---
-
-    no_print = [
-                'isTrain',
-                'df_source',
-                'mlp',
-                'net',
-                'input_list',
-                'label_list',
-                'period_name',
-                'mlp_num_inputs',
-                'num_outputs_for_label',
-                'datetime',
-                'device',
-                'weight_paths',
-                'train_save_datetime_dir',
-                'parameter_path',
-                'dataloader_params',
-                'model_params',
-                'train_conf_params',
-                'test_conf_params'
-                ]
-
-    if not params.isTrain:
-        no_print = no_print + ['augmentation', 'sampler', 'pretrained']
-
-    def _arg2str(param: str, arg: Union[str, int, float, None]) -> str:
-        """
-        Convert argument to string.
-
-        Args:
-            param (str): parameter
-            arg (Union[str, int, float, None): argument
-
-        Returns:
-            str: strings of argument
-        """
-        if param == 'lr':
-            if arg is None:
-                str_arg = 'Default'
-            else:
-                str_arg = str(param)
-            return str_arg
-        elif param == 'gpu_ids':
-            if arg == []:
-                str_arg = 'CPU selected'
-            else:
-                str_arg = f"{arg}  (Primary GPU:{arg[0]})"
-            return str_arg
-        else:
-            if arg is None:
-                str_arg = 'No need'
-            else:
-                str_arg = str(arg)
-            return str_arg
-
-    # Print
-    phase = 'Training' if params.isTrain else 'Test'
-    message = ''
-    message += f"{'-'*25} Options for {phase} {'-'*33}\n"
-
-    for _param, _arg in vars(params).items():
-        if _param not in no_print:
-            _str_arg = _arg2str(_param, _arg)
-            message += '{:>25}: {:<40}\n'.format(_param, _str_arg)
-
-    message += f"{'-'*30} End {'-'*48}\n"
-    logger.info(message)
 
 
 class BaseModel(ABC):
